feat(SpaceAvatar) move .js file to folder
