+++ src/components/SpaceList/SpaceAvatar/SpaceAvatar.js
@@ -7,10 +7,10 @@ import type { AvatarPlaceholder } from '@dlghq/dialog-types';
 
 import React, { PureComponent } from 'react';
 import classNames from 'classnames';
-import createSequence from '../../utils/createSequence';
-import getAvatarText from '../Avatar/utils/getAvatarText';
-import getAvatarColor from '../Avatar/utils/getAvatarColor';
-import styles from './SpaceList.css';
+import createSequence from '../../../utils/createSequence';
+import getAvatarText from '../../Avatar/utils/getAvatarText';
+import getAvatarColor from '../../Avatar/utils/getAvatarColor';
+import styles from '../SpaceList.css';
 
 export type Props = {
   id: string,
