+++ src/components/SpaceList/SpaceList.js
@@ -7,7 +7,7 @@ import type { AvatarPlaceholder } from '@dlghq/dialog-types';
 
 import React, { PureComponent } from 'react';
 import classNames from 'classnames';
-import SpaceAvatar from './SpaceAvatar.js';
+import SpaceAvatar from './SpaceAvatar/SpaceAvatar.js';
 import SpaceAdd from './SpaceAdd';
 import styles from './SpaceList.css';
